moved validate parameter value logic to validation to avoid duplication
